hal: Stream: wrap items and attach lifetimes
The problem at hand is known as the "stream iterator problem":

  Imagine a stream of items. Only one item shall be accessible at a time.
  Once the next item is acquired, the previous one is invalidated.

In Rust, we can model this by using explicit lifetimes: the next() call shall
output items with an attached lifetime which requires the stream itself to be
borrowed for as long as the item lives. As a consequence, one cannot fetch the
next stream element without dropping the current one.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>