hal: Device: Do not borrow self for the stream lifetime
This finally alleviates a huge issue faced by clients of this crate.
Many times, one just wants to capture images as part of a bigger application.
In such cases, the Device and Stream instances would often end up having to be
stored in the same struct, leading to self referential structs because the
stream borrows from device for its lifetime.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>