traits: Device: Tie returned stream lifetime to self
What we're saying is basically:
"Here's a stream which lives as long as the device it originated from".

The important thing to note here is that the lifetime of the stream is not
directly tied to to that of the device. Meaning you can create a stream
using a device and still access even mutable functions of the device while
the stream is alive.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>