control: Overhaul API
"The more things change, the more they stay the same".

Give new descriptive names to our types, such as Descriptor vs
Control and State vs Value.

Remove the Unknown type variant of the Descriptor enum. If there
is a type we cannot handle, you can either skip the platform
control and not expose it to the consumer API, or map it to a
String value variant.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>