Revert "Remove openpnp-capture HAL"
libuvc does not seem to support the latest macOS iSight webcams.
I tested this on my MacBook Pro 14" with the M1 Pro chip.

To restore functionality on recent Apple hardware, we bring back
the openpnp-capture library as a dependency for macOS.

libuvc will be used on Windows only going forward. Since maintaining
two separate backends for Windows and macOS is a considerable effort,
I might decide to axe the libuvc support completely in the future.

This reverts commit fbbb8a7da6008939fa69a01381dde7d8f18c9356.